refactor: change messages.js to operatorAdapter
