+++ lib/services/__tests__/account.test.js
@@ -3,7 +3,7 @@ import Config from 'react-native-config'
 import * as uuid from 'uuid'
 import * as crypto from '../crypto'
 import * as storage from '../storage'
-import * as messages from '../messages'
+import * as operatorAdapter from '../operatorAdapter'
 import { generateTestAccount } from './_helpers'
 
 jest.mock('uuid', () => ({
@@ -23,10 +23,10 @@ jest.mock('../storage', () => ({
     .mockName('storage.storeKey')
     .mockResolvedValue(),
 }))
-jest.mock('../messages', () => ({
+jest.mock('../operatorAdapter', () => ({
   createAccountRegistration: jest
     .fn()
-    .mockName('messages.createAccountRegistration'),
+    .mockName('operatorAdapter.createAccountRegistration'),
 }))
 
 Config.OPERATOR_URL = 'aTotallyLegitOperatorUrl'
@@ -67,7 +67,7 @@ describe('account', () => {
         privateKeyPem,
       })
       crypto.toPublicKey.mockReturnValue(publicKey)
-      messages.createAccountRegistration.mockResolvedValue(jwt)
+      operatorAdapter.createAccountRegistration.mockResolvedValue(jwt)
       fetch.mockResponse(JSON.stringify({ data: { id: 'abc123' } }))
     })
     describe('new user', () => {
@@ -95,7 +95,7 @@ describe('account', () => {
       })
       it('generates a token', async () => {
         const result = await accountService.save(account)
-        expect(messages.createAccountRegistration).toHaveBeenCalledWith(
+        expect(operatorAdapter.createAccountRegistration).toHaveBeenCalledWith(
           result,
           {
             publicKey,
