+++ lib/services/account.js
@@ -2,7 +2,7 @@ import Config from 'react-native-config'
 import { v4 } from 'uuid'
 import { generateKey, toPublicKey } from './crypto'
 import { storeAccount, storeKey } from './storage'
-import { createAccountRegistration } from './messages'
+import { createAccountRegistration } from './operatorAdapter'
 
 export async function save(account) {
   try {
