+++ lib/services/auth.js
@@ -1,13 +1,14 @@
 import { verify } from './jwt'
 import { getConnection, storeConnection, storeKey } from './storage'
 import Config from 'react-native-config'
+import { createConnectionInit } from './serviceAdapter'
+
 import {
-  createConnectionInit,
   createConnection,
   createConnectionResponse,
   createLogin,
   createLoginResponse,
-} from './messages'
+} from './operatorAdapter'
 import { v4 } from 'uuid'
 import { generateKey, toPublicKey } from './crypto'
