+++ lib/services/operatorAdapter.js
@@ -3,8 +3,6 @@ import { getAccount, getAccountKeys } from './storage'
 import Config from 'react-native-config'
 import { schemas } from '@egendata/messaging'
 
-const nowSeconds = () => Math.floor(Date.now() / 1000)
-
 export const createAccountRegistration = async (
   // eslint-disable-next-line camelcase
   { id, pds: { provider, access_token } },
@@ -25,26 +23,6 @@ export const createAccountRegistration = async (
       jwk: publicKey,
       alg: schemas.algs[0],
     },
-  )
-}
-
-export const createConnectionInit = async ({ aud, iss, sid }) => {
-  const { publicKey, privateKey, privateKeyPem } = await getAccountKeys()
-  const now = nowSeconds()
-  return sign(
-    {
-      type: 'CONNECTION_INIT',
-      aud: iss,
-      iss: aud,
-      sid,
-      iat: now,
-      exp: now + 60,
-    },
-    {
-      jwk: privateKey,
-      pem: privateKeyPem,
-    },
-    { jwk: publicKey, alg: schemas.algs[0] },
   )
 }
 
@@ -131,3 +109,11 @@ export const createLoginResponse = async loginPayload => {
     { jwk: publicKey, alg: schemas.algs[0] },
   )
 }
+
+// function postToOperator(body) {
+//   return fetch(Config.OPERATOR_URL, {
+//     method: 'POST',
+//     headers: { 'content-type': 'application/jwt' },
+//     body,
+//   })
+// }
